Move files into views/contribute directory
